Improve lightshow
Burst takes precedence unless the encoder value has changed. I find this slightly disorienting when trying to make parameter changes via the pots, but I think it's better overall. The problem with the pots is that incoming CV is indistinguishable from pot changes, so delaying burst display isn't practical (and if the distribution, div/mult or probably is modulated, it will be very strange).